@@ -1,4 +1,4 @@
-name: flutter_app
+name: fire_circle
 description: A new Flutter application.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 flutter:
   uses-material-design: true
@@ -18,6 +18,9 @@ flutter:
     - assets/images/started2.jpg
     - assets/images/started3.jpg
     - assets/images/welcome.jpg
+    - assets/images/camp_icon.png
+    - assets/images/osusowake.png
+    - assets/images/boad.png
 
 dependencies:
   flutter:
@@ -25,8 +28,11 @@ dependencies:
   cupertino_icons: ^1.0.2
   adobe_xd: ^1.0.0+1
   entity: ^2.0.0
-  firebase_core: ^0.3.0
-  cloud_firestore: ^0.9.5
+  firebase_core: ^1.1.0
+  cloud_firestore: ^1.0.7
+  flappy_search_bar: ^1.7.2
+  responsive_x: ^0.1.1
+  provider: ^5.0.0
 
 dev_dependencies:
   flutter_test:
@@ -34,5 +40,3 @@ dev_dependencies:
   cupertino_icons: ^1.0.2
   adobe_xd: ^1.0.0+1
   entity: ^2.0.0
-  firebase_core: ^0.3.0
-  cloud_firestore: ^0.9.5
