@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 
 class CampSite {
 
@@ -5,20 +6,29 @@ class CampSite {
 
   String name;
 
+  String postalCode;
+
   String address;
 
-  String siteImage;
+  AssetImage? siteImage;
 
-  CampSite({ this.campSiteId, this.name, this.address, this.siteImage});
+  int? checkInNum;
+
+  CampSite({ required this.campSiteId, required this.name,
+    required this.postalCode, required this.address, this.siteImage, this.checkInNum});
 
   factory CampSite.fromMap(Map<String, dynamic> map) =>
-      CampSite(campSiteId: map['camp_site_id'], name: map['name'], address: map['address'], siteImage: map['site_image']);
+      CampSite(campSiteId: map['camp_site_id'], name: map['name'],
+          postalCode: map['postal_code'], address: map['address'],
+          siteImage: map['site_image'], checkInNum: map['check_in_num']);
 
   Map<String, dynamic> toMap() => {
     'camp_site_id': campSiteId,
     'name': name,
+    'postal_code': postalCode,
     'address': address,
-    'site_image': siteImage
+    'site_image': siteImage,
+    'check_in_num': checkInNum
   };
 
 }
