@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:adobe_xd/pinned.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-class xd_checkin extends StatelessWidget {
-  xd_checkin({
-    Key key,
+class XdCheckin extends StatelessWidget {
+  XdCheckin({
+    Key? key,
   }) : super(key: key);
   @override
   Widget build(BuildContext context) {
